Renamed some files for better readability
