+++ server/app/controllers/movie.controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import Movie from "../models/movies.js";
+import Movie from "../models/movie.model.js";
 import { createMovie, getAllMovies, getMovieByName, updateMovie, deleteMovie } from "../service/movie.service.js";
 import { CreateMovieInput } from "../schema/movie.schema.js";
