+++ server/app/models/movie.model.ts
