+++ server/app/utils/swagger.ts
@@ -1,6 +1,6 @@
 import swaggerJsdoc from "swagger-jsdoc";
 import swaggerUi from "swagger-ui-express";
-import { Express } from "express"
+import { Express, Request, Response } from "express"
 // import version from "../../package.json" assert { type: "json"};
 const version = "1";
 
@@ -22,7 +22,7 @@ export function swaggerDocs(app: Express, port: number) {
   app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerSpec));
 
   // Docs in JSON format
-  app.get("/docs.json", (req, res) => {
+  app.get("/docs.json", (req: Request, res: Response) => {
     res.setHeader("Content-Type", "application/json");
     res.send(swaggerSpec);
   });
